refactor(onboarding): move coding agent onboarding enablement into shared onboarding provisioning
