+++ packages/backend/src/ee/services/ProjectService/provisionOnboardingHomepage.test.ts
@@ -11,6 +11,7 @@ import {
     type ProjectHomepage,
     type SessionUser,
 } from '@lightdash/common';
+import Logger from '../../../logging/logger';
 import {
     provisionOnboardingHomepage,
     type ProvisionOnboardingHomepageArguments,
@@ -158,11 +159,16 @@ describe('provisionOnboardingHomepage', () => {
 
         await provisionOnboardingHomepage(mocks.args);
 
-        expect(
-            mocks.ensureOrganizationOverrideEnabled,
-        ).toHaveBeenCalledExactlyOnceWith({
+        expect(mocks.ensureOrganizationOverrideEnabled).toHaveBeenCalledTimes(
+            2,
+        );
+        expect(mocks.ensureOrganizationOverrideEnabled).toHaveBeenCalledWith({
             user,
             featureFlagId: CommercialFeatureFlags.HomepageBuilder,
+        });
+        expect(mocks.ensureOrganizationOverrideEnabled).toHaveBeenCalledWith({
+            user,
+            featureFlagId: FeatureFlags.CodingAgentOnboarding,
         });
         expect(mocks.listHomepages).not.toHaveBeenCalled();
         expect(mocks.createHomepage).not.toHaveBeenCalled();
@@ -175,6 +181,7 @@ describe('provisionOnboardingHomepage', () => {
                 projectId: PROJECT_UUID,
                 onboardingFlow: 'new',
                 homepageBuilderEnablement: 'kept_disabled',
+                codingAgentOnboardingEnablement: 'kept_disabled',
                 reason: 'homepage_builder_flag_disabled',
             },
         });
@@ -202,6 +209,7 @@ describe('provisionOnboardingHomepage', () => {
                 projectId: PROJECT_UUID,
                 onboardingFlow: 'new',
                 homepageBuilderEnablement: 'failed',
+                codingAgentOnboardingEnablement: 'failed',
                 reason: 'homepage_builder_flag_disabled',
             },
         });
@@ -229,6 +237,7 @@ describe('provisionOnboardingHomepage', () => {
                 projectId: PROJECT_UUID,
                 onboardingFlow: 'legacy',
                 homepageBuilderEnablement: null,
+                codingAgentOnboardingEnablement: null,
                 reason: 'new_onboarding_flag_disabled',
             },
         });
@@ -250,6 +259,7 @@ describe('provisionOnboardingHomepage', () => {
                 projectId: PROJECT_UUID,
                 onboardingFlow: 'new',
                 homepageBuilderEnablement: 'enabled',
+                codingAgentOnboardingEnablement: 'enabled',
                 reason: 'homepage_already_exists',
             },
         });
@@ -276,6 +286,7 @@ describe('provisionOnboardingHomepage', () => {
                 projectId: PROJECT_UUID,
                 onboardingFlow: 'new',
                 homepageBuilderEnablement: null,
+                codingAgentOnboardingEnablement: null,
                 reason: 'not_first_project',
             },
         });
@@ -313,6 +324,7 @@ describe('provisionOnboardingHomepage', () => {
                 projectId: PROJECT_UUID,
                 onboardingFlow: 'new',
                 homepageBuilderEnablement: 'enabled',
+                codingAgentOnboardingEnablement: 'enabled',
                 errorType: 'Error',
             },
         });
@@ -334,6 +346,7 @@ describe('provisionOnboardingHomepage', () => {
                 projectId: PROJECT_UUID,
                 onboardingFlow: 'new',
                 homepageBuilderEnablement: 'enabled',
+                codingAgentOnboardingEnablement: 'enabled',
                 errorType: 'Error',
             },
         });
@@ -348,11 +361,16 @@ describe('provisionOnboardingHomepage', () => {
             user,
             featureFlagId: FeatureFlags.NewOnboarding,
         });
-        expect(
-            mocks.ensureOrganizationOverrideEnabled,
-        ).toHaveBeenCalledExactlyOnceWith({
+        expect(mocks.ensureOrganizationOverrideEnabled).toHaveBeenCalledTimes(
+            2,
+        );
+        expect(mocks.ensureOrganizationOverrideEnabled).toHaveBeenCalledWith({
             user,
             featureFlagId: CommercialFeatureFlags.HomepageBuilder,
+        });
+        expect(mocks.ensureOrganizationOverrideEnabled).toHaveBeenCalledWith({
+            user,
+            featureFlagId: FeatureFlags.CodingAgentOnboarding,
         });
         expect(mocks.getFeatureFlag).toHaveBeenCalledWith({
             user,
@@ -380,6 +398,7 @@ describe('provisionOnboardingHomepage', () => {
                 homepageUuid: HOMEPAGE_UUID,
                 onboardingFlow: 'new',
                 homepageBuilderEnablement: 'enabled',
+                codingAgentOnboardingEnablement: 'enabled',
             },
         });
     });
@@ -403,6 +422,71 @@ describe('provisionOnboardingHomepage', () => {
                 homepageUuid: HOMEPAGE_UUID,
                 onboardingFlow: 'new',
                 homepageBuilderEnablement: 'already_enabled',
+                codingAgentOnboardingEnablement: 'already_enabled',
+            },
+        });
+    });
+
+    it('provisions when coding agent onboarding enablement fails', async () => {
+        const errorSpy = vi
+            .spyOn(Logger, 'error')
+            .mockImplementation(() => Logger);
+        try {
+            const mocks = buildArguments();
+            mocks.ensureOrganizationOverrideEnabled.mockImplementation(
+                async ({ featureFlagId }) => {
+                    if (featureFlagId === FeatureFlags.CodingAgentOnboarding) {
+                        throw new Error('Enable failed');
+                    }
+                    return 'enabled';
+                },
+            );
+
+            await provisionOnboardingHomepage(mocks.args);
+
+            expect(mocks.createHomepage).toHaveBeenCalled();
+            expect(mocks.publishHomepage).toHaveBeenCalled();
+            expect(mocks.track).toHaveBeenCalledExactlyOnceWith({
+                event: 'onboarding_homepage.provisioned',
+                userId: USER_UUID,
+                properties: {
+                    organizationId: ORGANIZATION_UUID,
+                    projectId: PROJECT_UUID,
+                    homepageUuid: HOMEPAGE_UUID,
+                    onboardingFlow: 'new',
+                    homepageBuilderEnablement: 'enabled',
+                    codingAgentOnboardingEnablement: 'failed',
+                },
+            });
+            expect(errorSpy).toHaveBeenCalled();
+        } finally {
+            errorSpy.mockRestore();
+        }
+    });
+
+    it('provisions when coding agent onboarding remains disabled', async () => {
+        const mocks = buildArguments();
+        mocks.ensureOrganizationOverrideEnabled.mockImplementation(
+            async ({ featureFlagId }) =>
+                featureFlagId === FeatureFlags.CodingAgentOnboarding
+                    ? 'kept_disabled'
+                    : 'enabled',
+        );
+
+        await provisionOnboardingHomepage(mocks.args);
+
+        expect(mocks.createHomepage).toHaveBeenCalled();
+        expect(mocks.publishHomepage).toHaveBeenCalled();
+        expect(mocks.track).toHaveBeenCalledExactlyOnceWith({
+            event: 'onboarding_homepage.provisioned',
+            userId: USER_UUID,
+            properties: {
+                organizationId: ORGANIZATION_UUID,
+                projectId: PROJECT_UUID,
+                homepageUuid: HOMEPAGE_UUID,
+                onboardingFlow: 'new',
+                homepageBuilderEnablement: 'enabled',
+                codingAgentOnboardingEnablement: 'kept_disabled',
             },
         });
     });
