+++ packages/backend/src/ee/services/ProjectService/provisionOnboardingHomepage.ts
@@ -7,6 +7,7 @@ import {
 } from '@lightdash/common';
 import * as Sentry from '@sentry/node';
 import {
+    type CodingAgentOnboardingEnablement,
     type HomepageBuilderEnablement,
     type LightdashAnalytics,
     type OnboardingFlow,
@@ -55,6 +56,8 @@ export const provisionOnboardingHomepage = async ({
         ? 'new'
         : 'legacy';
     let homepageBuilderEnablement: HomepageBuilderEnablement | null = null;
+    let codingAgentOnboardingEnablement: CodingAgentOnboardingEnablement | null =
+        null;
     const trackSkipped = (reason: OnboardingHomepageSkippedReason) => {
         analytics.track({
             event: 'onboarding_homepage.skipped',
@@ -64,6 +67,7 @@ export const provisionOnboardingHomepage = async ({
                 projectId: projectUuid,
                 onboardingFlow,
                 homepageBuilderEnablement,
+                codingAgentOnboardingEnablement,
                 reason,
             },
         });
@@ -100,6 +104,22 @@ export const provisionOnboardingHomepage = async ({
             homepageBuilderEnablement = 'failed';
         }
 
+        try {
+            codingAgentOnboardingEnablement =
+                await featureFlagService.ensureOrganizationOverrideEnabled({
+                    user,
+                    featureFlagId: FeatureFlags.CodingAgentOnboarding,
+                });
+        } catch (error) {
+            Sentry.captureException(error);
+            Logger.error(
+                `Failed to enable coding agent onboarding for organization ${organizationUuid}: ${
+                    error instanceof Error ? error.message : String(error)
+                }`,
+            );
+            codingAgentOnboardingEnablement = 'failed';
+        }
+
         const homepageBuilderFlag = await featureFlagService.get({
             user,
             featureFlagId: CommercialFeatureFlags.HomepageBuilder,
@@ -133,6 +153,7 @@ export const provisionOnboardingHomepage = async ({
                 homepageUuid: homepage.homepageUuid,
                 onboardingFlow,
                 homepageBuilderEnablement,
+                codingAgentOnboardingEnablement,
             },
         });
     } catch (error) {
@@ -144,6 +165,7 @@ export const provisionOnboardingHomepage = async ({
                 projectId: projectUuid,
                 onboardingFlow,
                 homepageBuilderEnablement,
+                codingAgentOnboardingEnablement,
                 errorType: error instanceof Error ? error.name : 'Unknown',
             },
         });
